# services/identity-service/src/main/java/dev/mathalama/identityservice/application/dto/request/SignInRequest.java
package dev.mathalama.identityservice.application.dto.request;

import jakarta.validation.constraints.NotBlank;

public record SignInRequest(
        @NotBlank(message = "Login is required")
        String login,

        @NotBlank(message = "Password is required")
        String password
) {}