# services/identity-service/src/main/java/dev/mathalama/identityservice/presentation/advice/ErrorResponse.java
package dev.mathalama.identityservice.presentation.advice;

import java.time.Instant;

public record ErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        String path
) {}